SUBS-9 Braintree Drop In Improvements
* Added loading spinner/button disable to MG drop in to prevent double submissions
* Extracted Braintree drop in into a more reusable component
* Fixed loader/spinner placement issue on checkout
* Added payment priority options to bt drop in to show paypal first

SUBS-9
