+++ src/pages/Checkout/CheckoutPage.vue
@@ -76,7 +76,7 @@
 									@complete-transaction="completeTransaction"
 								/>
 
-								<drop-in-payment-wrapper
+								<checkout-drop-in-payment-wrapper
 									v-else-if="showDropInPayments"
 									:amount="creditNeeded"
 									@refreshtotals="refreshTotals"
@@ -184,8 +184,9 @@ import { settingEnabled } from '@/util/settingsUtils';
 import promoQuery from '@/graphql/query/promotionalBanner.graphql';
 import CheckoutHolidayPromo from '@/components/Checkout/CheckoutHolidayPromo';
 import PaymentWrapper from '@/components/Checkout/PaymentWrapper';
-import DropInPaymentWrapper from '@/components/Checkout/DropInPaymentWrapper';
+import CheckoutDropInPaymentWrapper from '@/components/Checkout/CheckoutDropInPaymentWrapper';
 import RandomLoanSelector from '@/components/RandomLoanSelector/randomLoanSelector';
+
 
 export default {
 	components: {
@@ -200,7 +201,7 @@ export default {
 		LoadingOverlay,
 		CheckoutHolidayPromo,
 		PaymentWrapper,
-		DropInPaymentWrapper,
+		CheckoutDropInPaymentWrapper,
 		RandomLoanSelector,
 	},
 	inject: ['apollo', 'kvAuth0'],
